tutorials/calendar-agent/pydantic-agent/pyproject.toml
[project]
name = "calendar-agent-pydantic"
version = "0.1.0"
description = "Calendar agent using Pydantic AI"
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai",
    "google-auth-oauthlib",
    "google-api-python-client",
    "anthropic",
]

[tool.uv]
dev-dependencies = [
    "pytest",
    "pytest-asyncio",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
